+++ instrumentation/azure-core/azure-core-1.36/javaagent/build.gradle.kts
@@ -6,7 +6,7 @@ muzzle {
   pass {
     group.set("com.azure")
     module.set("azure-core")
-    versions.set("[1.36.0,1.53.0)")
+    versions.set("[1.36.0,)")
     assertInverse.set(true)
   }
 }
@@ -29,10 +29,15 @@ dependencies {
   // Ensure no cross interference
   testInstrumentation(project(":instrumentation:azure-core:azure-core-1.14:javaagent"))
   testInstrumentation(project(":instrumentation:azure-core:azure-core-1.19:javaagent"))
-  testInstrumentation(project(":instrumentation:azure-core:azure-core-1.53:javaagent"))
 }
 
 val latestDepTest = findProperty("testLatestDeps") as Boolean
+
+tasks {
+  withType<Test>().configureEach {
+    systemProperty("testLatestDeps", findProperty("testLatestDeps") as Boolean)
+  }
+}
 
 testing {
   suites {
@@ -42,8 +47,10 @@ testing {
       dependencies {
         if (latestDepTest) {
           implementation("com.azure:azure-core:+")
+          implementation("com.azure:azure-core-test:+")
         } else {
           implementation("com.azure:azure-core:1.36.0")
+          implementation("com.azure:azure-core-test:1.16.2")
         }
       }
     }
