--- instrumentation/azure-core/azure-core-1.53/javaagent/src/main/resources/azure-core-1.53/META-INF/services/com.azure.core.util.tracing.TracerProvider
@@ -1,1 +1,0 @@
-io.opentelemetry.javaagent.instrumentation.azurecore.v1_53.shaded.com.azure.core.tracing.opentelemetry.OpenTelemetryTracerProvider
