+++ settings.gradle.kts
@@ -217,8 +217,6 @@ include(":instrumentation:azure-core:azure-core-1.19:javaagent")
 include(":instrumentation:azure-core:azure-core-1.19:library-instrumentation-shaded")
 include(":instrumentation:azure-core:azure-core-1.36:javaagent")
 include(":instrumentation:azure-core:azure-core-1.36:library-instrumentation-shaded")
-include(":instrumentation:azure-core:azure-core-1.53:javaagent")
-include(":instrumentation:azure-core:azure-core-1.53:library-instrumentation-shaded")
 include(":instrumentation:c3p0-0.9:javaagent")
 include(":instrumentation:c3p0-0.9:library")
 include(":instrumentation:c3p0-0.9:testing")
